+++ src/hooks/pinch-pan/use-pinch-session.ts
@@ -1,0 +1,29 @@
+import { useState } from 'react'
+
+export interface PinchSession {
+  referenceDistance: number
+  lastDistance: number
+}
+
+export function usePinchSession() {
+  const [pinchSession, setPinchSession] = useState<PinchSession | null>(null)
+
+  function setLastDistance(distance: number) {
+    setPinchSession((session) => {
+      if (!session) {
+        return null
+      }
+
+      return {
+        ...session,
+        lastDistance: distance,
+      }
+    })
+  }
+
+  return {
+    pinchSession,
+    setLastDistance,
+    setPinchSession,
+  }
+}
